Extract shared model setup to a base class
